afs-client-accessd: InactiveDestroy on DBI handles
We fork() whenever we spawn_exporter, in order to search for
exportable databases. When that child exits, currently that means it
will run the destructor for the sqlite db handle, since it has a copy
in memory, and we run destructors on objects when we exit(). This can
make the handle in the parent become unusable, failing with:

  DBD::SQLite::db commit failed: disk I/O error

when we try to commit in the parent. Confusingly, this only seems to
happen with certain (newer) versions of DBD::SQLite; presumably some
version added additional automatic "cleanup" in the destructor.

To fix this, set the InactiveDestroy setting on our DBI handles. This
means that the database will not be cleaned up when the handle goes
out of scope and gets destroyed. During normal operation we commit and
disconnect the db handle, so this should be fine. In the case of an
unclean shutdown or other interruption, we don't really care if the db
is inconsistent or does something else incorrect, since we only use
the collected data if there were no such errors when processing.