afs-vol-check: properly support multi-homed file servers
Run vos listaddrs to retrieve the uuid's of the file servers in the
cell. Use the file server uuid for the volume consistency checks,
instead of the ip addresses given by listvldb and listvol.

(I had originally assumed the vldb enties would always have the primary
address, but this is not case.)