afs-vol-check: check for old servers in vldb
Check for fileservers in the vldb which are not referenced by vl
entries.  This happens when a fileserver is decommissioned and
the admin did not remove the mh entry with vos change addr.

Also, fix the vos listaddrs command options; use -printuuid
instead of --printuuid. The latter is not supported by old
versions of vos.