collect rx and xstats from fileservers
This is a simple front-end to rxdebug and xstat_fs_test to collect
stats from the file servers in one or more cells.  This tool requires
patched versions of rxdebug and xstat_fs_test for now, since it uses
some new output format options to save the data in a more machine
readable format.

The servers to be accessed are specified in a config file.  The config
file is created the first time this script is run.

When the config file is missing, the script will attempt to find the
local cell name with fs, and the list of file server addresses with vos
listaddrs.  If you want to skip this auto-detection, explicitly specify
the addresses in the config file.