afs-client-accessd: allow configuration of export time
Currently, every single host running afs-client-accessd exports the
collected data at midnight. Depending on how many hosts we have,
exporting several databases in parallel can overload the central host
and, as a result of that, substantially increase the time needed to
complete this operation.

To mitigate this problem, introduce a new configuration directive that
can be used to change the hour in which the databases should be
exported. With this directive, the number of nodes transferring data to
the central host at the same time can be reduced by setting different
export times for different hosts.